requirements.txt
fastapi==0.111.0
pandas==2.2.2
numpy==1.26.4
psycopg2-binary==2.9.9
python-socketio==5.11.2
websocket-client==1.8.0
pyaml-env==1.2.1
python-decouple==3.8
uvicorn==0.29.0
pydantic==1.10.14
sqlalchemy